Fix dialog i18n not change after set language issue.
